+++ src/app/app.component.html
@@ -38,10 +38,10 @@
     <a
       mat-menu-item
       [routerLinkActive]="['active']"
-      [routerLink]="['/archive-settings/']"
+      [routerLink]="['/policies/']"
     >
       <i class="material-icons"> cloud_download </i>
-      Archive Settings</a
+      Policies</a
     >
     <a mat-menu-item [routerLinkActive]="['active']" [routerLink]="['/samples/']" >
       <i class="material-icons"> center_focus_weak </i>
