--- src/app/policies/archive-settings/archive-settings.component.html
@@ -1,83 +1,0 @@
-<div class="archive-settings">
-</div>
-
-
-  <div class="button-group">
-    <ng-container *ngIf="editEnabled" >
-      <button mat-button mat-button (click)="onEditClick()" color="primary">
-        <mat-icon >edit</mat-icon> Edit Selection
-    </button>
-
-    </ng-container>
-  </div>
-
-  <mat-paginator class="policy-paginator" [pageSizeOptions]="pageSizeOptions" [pageIndex]="currentPage" [pageSize]="policiesPerPage" [length]="policyCount$ | async" (page)="onPageChange($event)">
-  </mat-paginator>
-
-  <mat-table [dataSource]="policies$ | async" class="policy-table mat-elevation-z2" matSort (matSortChange)="onSortChange($event)">
-
-    <!-- Column Definitions -->
-    <ng-container matColumnDef="select">
-      <mat-header-cell *matHeaderCellDef>
-        <mat-checkbox [checked]="allAreSelected()" (change)="handleSelectAll($event)">
-        </mat-checkbox>
-      </mat-header-cell>
-      <mat-cell *matCellDef="let row">
-        <mat-checkbox [checked]="isChecked(row)" (change)="handleSelect($event, row)" (click)="$event.stopPropagation()">
-        </mat-checkbox>
-      </mat-cell>
-    </ng-container>
-
-
-
-    <ng-container matColumnDef="manager">
-      <mat-header-cell *matHeaderCellDef mat-sort-header>Manager</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.manager}}</mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="tapeRedundancy">
-      <mat-header-cell *matHeaderCellDef>Tape Redundancy</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.tapeRedundancy}}</mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="autoArchive">
-      <mat-header-cell *matHeaderCellDef>Auto Archive</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.autoArchive}}</mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="autoArchiveDelay">
-      <mat-header-cell *matHeaderCellDef mat-sort-header>Auto Archive Delay</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.autoArchiveDelay}}</mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="ownerGroup">
-      <mat-header-cell *matHeaderCellDef>Group</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.ownerGroup}}</mat-cell>
-    </ng-container>
-
-
-
-    <ng-container matColumnDef="archiveEmailNotification">
-      <mat-header-cell *matHeaderCellDef mat-sort-header>Archive Email Notification</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.archiveEmailNotification }}</mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="archiveEmailsToBeNotified">
-      <mat-header-cell *matHeaderCellDef>Archive Emails to be Notified</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.archiveEmailsToBeNotified}}</mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="retrieveEmailNotification">
-      <mat-header-cell *matHeaderCellDef>Retrieve Email Notification</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.retrieveEmailNotification}}</mat-cell>
-    </ng-container>
-    <ng-container matColumnDef="retrieveEmailsToBeNotified">
-      <mat-header-cell *matHeaderCellDef>Retrieve Emails to be Notified</mat-header-cell>
-      <mat-cell *matCellDef="let policy">{{policy.retrieveEmailsToBeNotified}}</mat-cell>
-    </ng-container>
-
-
-    <!-- End of Column Definitions -->
-
-    <mat-header-row class='policies-header' *matHeaderRowDef="getDisplayedColumns()"></mat-header-row>
-    <mat-row class='row-generic' *matRowDef="let policy; columns: getDisplayedColumns()" (click)="handleClick(policy)">
-    </mat-row>
-  </mat-table>
