+++ src/app/policies/policies/policies.component.scss
@@ -63,7 +63,7 @@ mat-cell, .mat-cell {
   cursor: pointer;
 }
 
-/*:host /deep/*/ .archive-settings-header {
+/*:host /deep/*/ .policies-header {
   background-color: mat-color($primary);
 }
